day 5
WHY IS IT SO HARD AND IT'S ONLY DAY 5...

Used a bruteforce approach for part 1 at first and then saw the crazy input file so I had to come up with a different solution. Not that big of a deal.
PART 2? Used a bruteforce approach at first as well. I ran it for a couple of minutes and the python script was already taking 11GB of ram and still rising because it was trying to create a dictionary with over 15 billion numbers... Then I struggled with coming up and implementing a better solution but with the help of analyzing a dozen of other people's solutions I managed to do it...