day 8
this one was quite nice

EXCEPT THE INSTRUCTIONS. THEY COULD AND SHOULD'VE BEEN CLEARER.

I understood that the first node is in the first line in input, and last node is in the last line, because both examples in part 1 had them in first and last line. So it took me a good hour of debugging and then manually going through a big part of input nodes to figure out it has to start at node 'AAA' specifically and NOT at the node in the first line. Ughghghghghhhhh. I was so confused because everything looked right. It went from the first node all the way to the last one without any mistakes so I couldn't figure out why my answer was incorrect. Then when I went to manually check nodes in the input, I came up across 'AAA' node and the thought that it should've started at that node and not node in line one jumped in my mind.