day 9
I really liked this one.
Did a mistake that costed me an hour to fix but other than that it worked as planned.
What was the mistake?
My big brain though that using sum(epic_list) == 0 to check if the list contains only zeroes would be a good solution. Well... it most certainly was not: sum[2, 1, 0, -1, -2]) == 0...