day 5
The problem was so confusing I actually had to stop and THINK how I'll solve it. Dictionary immediately came to mind and soon after, a solution, which I couldn't fully render in my brain. I just felt like it should work, but couldn't really explain why. So I wrote the code. It went buttery smooth. I ran it. It worked on the first try. WHAT KIND OF MAGIC IS THIS THING CALLED "thinking". It's overpowered. Should be nerfed.